Revise import statements
